refactor(checker): relocate per-file memo type aliases to context/aliases.rs (2000-line cap)
Move AccessorLevelsCacheValue, AccessorLevelsCache, MemberAccessInfoCache,
and CallbackMismatchMemo out of context/mod.rs into the existing
context/aliases.rs submodule (re-exported via `pub use aliases::*`, so all
`crate::context::*` paths are unchanged). This brings mod.rs from 2018 back
to 1987 physical lines, under the checker-boundary 2000-LOC arch_guard cap
that the #13250 callback-mismatch memo field pushed it over.

Behavior unchanged: only type-alias definitions moved; the memo field, its
constructor init, file-session reset, and lifetime registration stay in place.
